backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "atlas-provider-sqlalchemy>=0.2.4",
    "fastapi>=0.115.12",
    "greenlet>=3.2.2",
    "httpx>=0.28.1",
    "openai>=1.82.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "spotipy>=2.25.1",
    "sqlmodel>=0.0.24",
    "testcontainers[postgres,postgresql]>=4.10.0",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.11",
]

[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
lint.ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
lint.isort.known-first-party = ["unwrapped"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
